pyproject.toml
[project]
name = "autonomous-browser-ai-agent-bedrock"
version = "0.1.0"
description = "Autonomous browser agent built on Playwright and LLM planning"
readme = "README.md"
authors = [
    { name = "Kaangml", email = "gumelekaan@gmail.com" }
]
requires-python = ">=3.10,<3.13"
dependencies = [
    "playwright>=1.48.0,<2.0.0",
    "pytest>=8.2.0,<9.0.0",
    "pytest-asyncio>=0.23.6,<0.24.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.3.0",
    "langchain-aws>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent", "src/browser", "src/config", "src/controller", "src/examples", "src/llm"]

[tool.hatch.build.targets.sdist]
include = [
    "src",
    "README.md",
    "pyproject.toml",
]
